+++ src/main/java/com/capstone/carbonlive/entity/Electricity.java
@@ -1,27 +1,26 @@
-package com.capstone.carbonlive.carbon;
+package com.capstone.carbonlive.entity;
 
-import com.capstone.carbonlive.building.Building;
+import com.capstone.carbonlive.entity.Building;
 import jakarta.persistence.*;
 import lombok.*;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Entity
 @Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @ToString
-public class Carbon {
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Electricity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "CARBON_ID")
+    @Column(name = "ELECTRICITY_ID")
     private Long id;
 
     @Column(nullable = false)
     private LocalDate recordedAt;
 
-    @Column(nullable = false, precision = 10, scale = 2)
-    private BigDecimal usages;
+    @Column(nullable = false)
+    private int usages;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "BUILDING_ID", nullable = false)
@@ -29,7 +28,7 @@ public class Carbon {
     private Building building;
 
     @Builder(toBuilder = true)
-    public Carbon(LocalDate recordedAt, BigDecimal usages, Building building) {
+    public Electricity(LocalDate recordedAt, int usages, Building building) {
         this.recordedAt = recordedAt;
         this.usages = usages;
         this.building = building;
