+++ src/main/java/com/capstone/carbonlive/repository/BuildingRepository.java
@@ -1,6 +1,6 @@
-package com.capstone.carbonlive.building.repository;
+package com.capstone.carbonlive.repository;
 
-import com.capstone.carbonlive.building.Building;
+import com.capstone.carbonlive.entity.Building;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BuildingRepository extends JpaRepository<Building, Long> {
